+++ cypress/plugins/index.js
@@ -13,7 +13,13 @@
 require('dotenv').config();
 const { addMatchImageSnapshotPlugin } = require('cypress-image-snapshot/plugin');
 
-const { setupGitHub, teardownGitHub, setupGitHubTest, teardownGitHubTest } = require('./github');
+const {
+  setupGitHub,
+  teardownGitHub,
+  setupGitHubTest,
+  teardownGitHubTest,
+  seedGitHubRepo,
+} = require('./github');
 const {
   setupGitGateway,
   teardownGitGateway,
@@ -29,7 +35,7 @@ const {
 } = require('./bitbucket');
 const { setupProxy, teardownProxy, setupProxyTest, teardownProxyTest } = require('./proxy');
 
-const { copyBackendFiles } = require('../utils/config');
+const { copyBackendFiles, switchVersion } = require('../utils/config');
 
 module.exports = async (on, config) => {
   // `on` is used to hook into various events Cypress emits
@@ -135,20 +141,42 @@ module.exports = async (on, config) => {
 
       return null;
     },
-  });
+    async seedRepo(taskData) {
+      const { backend } = taskData;
 
-  on('before:browser:launch', (browser = {}, args) => {
-    if (browser.name === 'chrome') {
-      // to allows usage of a mock proxy
-      args.push('--ignore-certificate-errors');
-      args.push('-–disable-gpu');
-      if (browser.isHeaded) {
-        args.push('--window-size=1200,1200');
-      } else {
-        args.push('--window-size=1200,1077');
+      console.log(`Seeding repository for backend`, backend);
+
+      switch (backend) {
+        case 'github':
+          await seedGitHubRepo(taskData);
+          break;
       }
 
-      return args;
+      return null;
+    },
+    async switchToVersion(taskData) {
+      const { version } = taskData;
+
+      console.log(`Switching CMS to version '${version}'`);
+
+      await switchVersion(version);
+
+      return null;
+    },
+  });
+
+  on('before:browser:launch', (browser = {}, launchOptions) => {
+    if (browser.name === 'chrome') {
+      // to allows usage of a mock proxy
+      launchOptions.args.push('--ignore-certificate-errors');
+      launchOptions.args.push('-–disable-gpu');
+      if (browser.isHeaded) {
+        launchOptions.args.push('--window-size=1200,1200');
+      } else {
+        launchOptions.args.push('--window-size=1200,1077');
+      }
+
+      return launchOptions;
     }
   });