+++ packages/netlify-cms-backend-git-gateway/src/GitHubAPI.ts
@@ -71,6 +71,20 @@ export default class API extends GithubAPI {
     return Promise.resolve({ login: '', ...this.commitAuthor });
   }
 
+  async getHeadReference(head: string) {
+    if (!this.repoOwner) {
+      // get the repo owner from the branch url
+      // this is required for returning the full head reference, e.g. owner:head
+      // when filtering pull requests based on the head
+      const branch = await this.getDefaultBranch();
+      const self = branch._links.self;
+      const regex = new RegExp('https?://.+?/repos/(.+?)/');
+      const owner = self.match(regex);
+      this.repoOwner = owner ? owner[1] : '';
+    }
+    return super.getHeadReference(head);
+  }
+
   commit(message: string, changeTree: { parentSha?: string; sha: string }) {
     const commitParams: {
       message: string;