+++ packages/netlify-cms-backend-github/src/fragments.ts
@@ -3,6 +3,7 @@ import gql from 'graphql-tag';
 export const repository = gql`
   fragment RepositoryParts on Repository {
     id
+    isFork
   }
 `;
 
@@ -41,6 +42,7 @@ export const pullRequest = gql`
   fragment PullRequestParts on PullRequest {
     id
     baseRefName
+    baseRefOid
     body
     headRefName
     headRefOid
@@ -50,6 +52,11 @@ export const pullRequest = gql`
     merged_at: mergedAt
     repository {
       ...RepositoryParts
+    }
+    labels(last: 100) {
+      nodes {
+        name
+      }
     }
   }
   ${repository}