+++ packages/netlify-cms-core/src/components/Collection/Entries/EntryCard.js
@@ -1,11 +1,13 @@
 import React from 'react';
 import styled from '@emotion/styled';
 import { connect } from 'react-redux';
-import { getAsset } from 'Actions/media';
+import { boundGetAsset } from 'Actions/media';
 import { Link } from 'react-router-dom';
-import { colors, colorsRaw, components, lengths, Asset } from 'netlify-cms-ui-legacy';
+import { colors, colorsRaw, components, lengths } from 'netlify-cms-ui-legacy';
 import { VIEW_STYLE_LIST, VIEW_STYLE_GRID } from 'Constants/collectionViews';
 import { summaryFormatter } from 'Lib/formatters';
+import { keyToPathArray } from 'Lib/stringTemplate';
+import { selectIsLoadingAsset } from 'Reducers/medias';
 
 const ListCard = styled.li`
   ${components.card};
@@ -76,25 +78,21 @@ const CardBody = styled.div`
 `;
 
 const CardImage = styled.div`
-  background-image: url(${props => props.value?.toString()});
+  background-image: url(${props => props.src});
   background-position: center center;
   background-size: cover;
   background-repeat: no-repeat;
   height: 150px;
 `;
 
-const CardImageAsset = ({ getAsset, image, folder }) => {
-  return <Asset folder={folder} path={image} getAsset={getAsset} component={CardImage} />;
-};
-
 const EntryCard = ({
   path,
   summary,
   image,
-  imageFolder,
+  imageField,
   collectionLabel,
   viewStyle = VIEW_STYLE_LIST,
-  boundGetAsset,
+  getAsset,
 }) => {
   if (viewStyle === VIEW_STYLE_LIST) {
     return (
@@ -107,6 +105,9 @@ const EntryCard = ({
     );
   }
 
+  const asset = getAsset(image, imageField);
+  const src = asset.toString();
+
   if (viewStyle === VIEW_STYLE_GRID) {
     return (
       <GridCard>
@@ -115,9 +116,7 @@ const EntryCard = ({
             {collectionLabel ? <CollectionLabel>{collectionLabel}</CollectionLabel> : null}
             <CardHeading>{summary}</CardHeading>
           </CardBody>
-          {image ? (
-            <CardImageAsset getAsset={boundGetAsset} image={image} folder={imageFolder} />
-          ) : null}
+          {image ? <CardImage src={src} /> : null}
         </GridCardLink>
       </GridCard>
     );
@@ -128,7 +127,9 @@ const mapStateToProps = (state, ownProps) => {
   const { entry, inferedFields, collection } = ownProps;
   const label = entry.get('label');
   const entryData = entry.get('data');
-  const defaultTitle = label || entryData.get(inferedFields.titleField);
+  const defaultTitle =
+    label ||
+    (inferedFields.titleField && entryData.getIn(keyToPathArray(inferedFields.titleField)));
   const summaryTemplate = collection.get('summary');
   const summary = summaryTemplate
     ? summaryFormatter(summaryTemplate, entry, collection)
@@ -139,21 +140,23 @@ const mapStateToProps = (state, ownProps) => {
     image = encodeURI(image);
   }
 
+  const isLoadingAsset = selectIsLoadingAsset(state.medias);
+
   return {
     summary,
     path: `/collections/${collection.get('name')}/entries/${entry.get('slug')}`,
     image,
     imageFolder: collection
       .get('fields')
-      ?.find(f => f.get('name') === inferedFields.imageField && f.get('widget') === 'image')
-      ?.get('media_folder'),
+      ?.find(f => f.get('name') === inferedFields.imageField && f.get('widget') === 'image'),
+    isLoadingAsset,
   };
 };
 
-const mapDispatchToProps = {
-  boundGetAsset: (collection, entry) => (dispatch, getState) => (path, folder) => {
-    return getAsset({ collection, entry, path, folder })(dispatch, getState);
-  },
+const mapDispatchToProps = dispatch => {
+  return {
+    boundGetAsset: (collection, entry) => boundGetAsset(dispatch, collection, entry),
+  };
 };
 
 const mergeProps = (stateProps, dispatchProps, ownProps) => {
@@ -161,7 +164,7 @@ const mergeProps = (stateProps, dispatchProps, ownProps) => {
     ...stateProps,
     ...dispatchProps,
     ...ownProps,
-    boundGetAsset: dispatchProps.boundGetAsset(ownProps.collection, ownProps.entry),
+    getAsset: dispatchProps.boundGetAsset(ownProps.collection, ownProps.entry),
   };
 };