+++ packages/netlify-cms-proxy-server/CHANGELOG.md
@@ -3,6 +3,73 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.1.5](https://github.com/netlify/netlify-cms/tree/master/packages/netlify-cms-proxy-server/compare/netlify-cms-proxy-server@1.1.4...netlify-cms-proxy-server@1.1.5) (2020-02-25)
+
+**Note:** Version bump only for package netlify-cms-proxy-server
+
+
+
+
+
+## [1.1.4](https://github.com/netlify/netlify-cms/tree/master/packages/netlify-cms-proxy-server/compare/netlify-cms-proxy-server@1.1.3...netlify-cms-proxy-server@1.1.4) (2020-02-17)
+
+**Note:** Version bump only for package netlify-cms-proxy-server
+
+
+
+
+
+## [1.1.3](https://github.com/netlify/netlify-cms/tree/master/packages/netlify-cms-proxy-server/compare/netlify-cms-proxy-server@1.1.2...netlify-cms-proxy-server@1.1.3) (2020-02-14)
+
+
+### Bug Fixes
+
+* **proxy-server:** add missing labels to file collections ([#3250](https://github.com/netlify/netlify-cms/tree/master/packages/netlify-cms-proxy-server/issues/3250)) ([8d67de0](https://github.com/netlify/netlify-cms/tree/master/packages/netlify-cms-proxy-server/commit/8d67de0e681e02ddbf811eecc5e4744b324b7e61))
+
+
+
+
+
+## [1.1.2](https://github.com/netlify/netlify-cms/tree/master/packages/netlify-cms-proxy-server/compare/netlify-cms-proxy-server@1.1.1...netlify-cms-proxy-server@1.1.2) (2020-02-12)
+
+
+### Bug Fixes
+
+* **proxy-server:** better handle files based collections ([#3237](https://github.com/netlify/netlify-cms/tree/master/packages/netlify-cms-proxy-server/issues/3237)) ([3678053](https://github.com/netlify/netlify-cms/tree/master/packages/netlify-cms-proxy-server/commit/3678053f0c419718000a21768790618749df762c))
+
+
+
+
+
+## [1.1.1](https://github.com/netlify/netlify-cms/tree/master/packages/netlify-cms-proxy-server/compare/netlify-cms-proxy-server@1.1.0...netlify-cms-proxy-server@1.1.1) (2020-02-11)
+
+**Note:** Version bump only for package netlify-cms-proxy-server
+
+
+
+
+
+# [1.1.0](https://github.com/netlify/netlify-cms/tree/master/packages/netlify-cms-proxy-server/compare/netlify-cms-proxy-server@1.0.7...netlify-cms-proxy-server@1.1.0) (2020-02-10)
+
+
+### Bug Fixes
+
+* fs proxy info action ([#3225](https://github.com/netlify/netlify-cms/tree/master/packages/netlify-cms-proxy-server/issues/3225)) ([4522739](https://github.com/netlify/netlify-cms/tree/master/packages/netlify-cms-proxy-server/commit/45227392315f4b8b7566ffe65756ca13dc55b9f8))
+
+
+### Features
+
+* **proxy-server:** add local fs middleware and make it the default ([#3217](https://github.com/netlify/netlify-cms/tree/master/packages/netlify-cms-proxy-server/issues/3217)) ([31dbd72](https://github.com/netlify/netlify-cms/tree/master/packages/netlify-cms-proxy-server/commit/31dbd72273b723bb6bbb551641a6e4bcc1f0314b))
+
+
+### Reverts
+
+* Revert "chore(release): publish" ([a015d1d](https://github.com/netlify/netlify-cms/tree/master/packages/netlify-cms-proxy-server/commit/a015d1d92a4b1c0130c44fcef1c9ecdb157a0f07))
+
+
+
+
+
 ## [1.0.7](https://github.com/netlify/netlify-cms/tree/master/packages/netlify-cms-proxy-server/compare/netlify-cms-proxy-server@1.0.6...netlify-cms-proxy-server@1.0.7) (2020-02-06)
 
 **Note:** Version bump only for package netlify-cms-proxy-server